Split app.py into route modules (#32)
* Split server/app.py into route modules

* add prefixes to API routers

* fix: address PR review comments
